mvp project structure, start moving view logic to presenter
