+++ app/src/main/java/com/example/sprintone/data/model/Gallery.java
@@ -1,0 +1,7 @@
+package com.example.sprintone.data.model;
+
+import com.example.sprintone.gallery.GalleryActivityMVP;
+
+public class Gallery implements GalleryActivityMVP.Model {
+
+}
