+++ app/src/main/java/com/example/sprintone/gallery/GalleryActivityMVP.java
@@ -1,0 +1,29 @@
+package com.example.sprintone.gallery;
+
+import com.example.sprintone.traversal.GallerySingleton;
+
+public interface GalleryActivityMVP {
+
+    interface Model {
+
+    }
+
+    interface View {
+
+    }
+
+    interface Presenter {
+        void setView(GalleryActivityMVP.View view);
+
+        void addPhoto();
+
+        int deletePhoto(GallerySingleton gallery);
+
+        void editCaption(GallerySingleton gallery, String edit_caption);
+
+
+    }
+
+
+
+}
